refactor : divide question apis #48
